pyproject.toml
[tool.poetry]
name = "optimus-plex"
version = "1.0.0"
description = "Optimise your Plex media Library"
authors = ["Pablo Minue"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ffmpeg-python = "^0.2.0"
pandas = "^2.2.3"
rich = "^13.9.4"
plexapi = {extras = ["alert"], version = "^4.16.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
